Stop importing from nuxt.config.js
