+++ nuxt.config.js
@@ -1,24 +1,8 @@
 import pkg from './package.json'
 import locales from './src/locales/scripts/valid-locales.json'
-import stringToBoolean from './src/utils/string-to-boolean'
 import { VIEWPORTS } from './src/constants/screens'
-
-/**
- * Default environment variables are set on this key. Defaults are fallbacks to existing env vars.
- * All boolean values should be designed to be false by default.
- */
-export const env = {
-  apiUrl: process.env.API_URL ?? 'https://api.openverse.engineering/v1/',
-  enableGoogleAnalytics: stringToBoolean(process.env.ENABLE_GOOGLE_ANALYTICS),
-  googleAnalyticsUA: process.env.GOOGLE_ANALYTICS_UA ?? 'UA-2010376-36',
-  filterStorageKey: 'openverse-filter-visibility',
-  notificationStorageKey: 'openverse-show-notification',
-  enableInternalAnalytics: stringToBoolean(
-    process.env.ENABLE_INTERNAL_ANALYTICS
-  ),
-  /** Feature flag to enable non-image media */
-  enableAudio: stringToBoolean(process.env.ENABLE_AUDIO),
-}
+import { dev } from './src/utils/dev'
+import { env } from './src/utils/env'
 
 /**
  * The default metadata for the site. Can be extended and/or overwritten per page. And even in components!
@@ -154,7 +138,7 @@ export default {
   ],
   head,
   env,
-  dev: process.env.NODE_ENV !== 'production',
+  dev,
   buildModules: [
     '@nuxtjs/composition-api/module',
     '@nuxt/postcss8',
