+++ src/mixins/iframe-height.js
@@ -1,6 +1,5 @@
 import debounce from 'lodash.debounce'
 import { sendWindowMessage } from '~/utils/send-message'
-import config from '../../nuxt.config.js'
 
 /**
  * When the app is in embedded mode, it passes the full height
@@ -43,7 +42,6 @@ export default {
     },
     notifyOuterWindow(height) {
       sendWindowMessage({
-        debug: config.dev,
         type: 'resize',
         value: { height },
       })
