+++ src/plugins/url-change.js
@@ -1,5 +1,5 @@
 import { sendWindowMessage } from '~/utils/send-message'
-import config from '../../nuxt.config.js'
+
 /**
  * In embedded mode, we need to notify the outer window of the current URL.
  * Normally, the `src/middleware/embed.js` middleware does this on every
@@ -10,7 +10,6 @@ export default function ({ app, store }) {
   app.router.onReady(() => {
     if (process.client && store.state.nav.isEmbedded) {
       sendWindowMessage({
-        debug: config.dev,
         type: 'urlChange',
         value: {
           path: app.router.currentRoute.fullPath,
