+++ src/utils/dev.js
@@ -1,0 +1,1 @@
+export const dev = process.env.NODE_ENV !== 'production'
